# openapi.yaml
openapi: 3.0.3
info:
  title: Go Messaging Bot API
  description: |
    Universal notification system with Telegram bot integration, admin approval system, and button-based user interface.
    
    ## Features
    - 🤖 Telegram Bot with inline keyboards
    - 👑 Admin approval workflow  
    - 🔐 HTTP Basic Authentication
    - 📱 Button-based UI (no command typing)
    - ⏰ Auto-cleanup of pending users
    - 📊 Real-time statistics
    
    ## Authentication
    Admin endpoints require HTTP Basic Authentication:
    - Username: `admin`
    - Password: `admin123` (change in production)
    
    ## Getting Started
    1. Create admin user via `/api/v1/admin/create`
    2. Users register via Telegram bot `/start`
    3. Admins approve users via Telegram `/admin` or API
    
  version: 2.0.0
  contact:
    name: Go Messaging Bot Support
    url: https://github.com/your-repo/go-messaging
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://your-domain.com/api/v1
    description: Production server
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Authentication for admin endpoints
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        telegram_user_id:
          type: integer
          format: int64
          description: Telegram user ID
        username:
          type: string
          nullable: true
          description: Telegram username
        first_name:
          type: string
          nullable: true
          description: User's first name
        last_name:
          type: string
          nullable: true
          description: User's last name
        language_code:
          type: string
          nullable: true
          description: User's language code
        is_bot:
          type: boolean
          description: Whether the user is a bot
        role:
          type: string
          enum: [user, admin]
          description: User role
        approval_status:
          type: string
          enum: [pending, approved, rejected, disabled]
          description: User approval status
        approved_by:
          type: string
          format: uuid
          nullable: true
          description: ID of admin who approved the user
        approved_at:
          type: string
          format: date-time
          nullable: true
          description: When the user was approved
        created_at:
          type: string
          format: date-time
          description: When the user was created
        updated_at:
          type: string
          format: date-time
          description: When the user was last updated
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        details:
          type: string
          description: Detailed error information
...
  parameters:
    UserIdPath:
      name: userID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User ID
...
paths:
...
  /users:
    get:
      summary: List all users
      description: Retrieve a paginated list of all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{id}:
    get:
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Admin Management Endpoints
...